Combine RTP Packet and PacketParser, polish RTCP
